image-grid/requirements.txt
fastapi
uvicorn
mangum
boto3
python-multipart
pillow
